+++ firebase-dataconnect/src/test/kotlin/com/google/firebase/dataconnect/CacheSettingsUnitTest.kt
@@ -38,7 +38,6 @@ import io.kotest.property.PropTestConfig
 import io.kotest.property.arbitrary.choice
 import io.kotest.property.arbitrary.enum
 import io.kotest.property.arbitrary.int
-import io.kotest.property.arbitrary.map
 import io.kotest.property.arbitrary.shuffle
 import io.kotest.property.arbitrary.string
 import io.kotest.property.checkAll
@@ -65,7 +64,7 @@ class CacheSettingsUnitTest {
 
   @Test
   fun `toString() returns a string that incorporates all property values`() = runTest {
-    checkAll(propTestConfig, cacheSettingsArb()) { cacheSettings: CacheSettings ->
+    checkAll(propTestConfig, Arb.dataConnect.cacheSettings()) { cacheSettings: CacheSettings ->
       val toStringResult = cacheSettings.toString()
       assertSoftly {
         toStringResult shouldStartWith "CacheSettings("
@@ -77,14 +76,14 @@ class CacheSettingsUnitTest {
 
   @Test
   fun `equals() should return true for the exact same instance`() = runTest {
-    checkAll(propTestConfig, cacheSettingsArb()) { cacheSettings: CacheSettings ->
+    checkAll(propTestConfig, Arb.dataConnect.cacheSettings()) { cacheSettings: CacheSettings ->
       cacheSettings.equals(cacheSettings) shouldBe true
     }
   }
 
   @Test
   fun `equals() should return true for an equal instance`() = runTest {
-    checkAll(propTestConfig, cacheSettingsArb()) { cacheSettings1: CacheSettings ->
+    checkAll(propTestConfig, Arb.dataConnect.cacheSettings()) { cacheSettings1: CacheSettings ->
       val cacheSettings2 = cacheSettings1.copy()
       withClue("cacheSettings1=$cacheSettings1 cacheSettings2=$cacheSettings2") {
         cacheSettings1.equals(cacheSettings2) shouldBe true
@@ -94,7 +93,7 @@ class CacheSettingsUnitTest {
 
   @Test
   fun `equals() should return false for null`() = runTest {
-    checkAll(propTestConfig, cacheSettingsArb()) { cacheSettings: CacheSettings ->
+    checkAll(propTestConfig, Arb.dataConnect.cacheSettings()) { cacheSettings: CacheSettings ->
       cacheSettings.equals(null) shouldBe false
     }
   }
@@ -108,7 +107,7 @@ class CacheSettingsUnitTest {
         Arb.dataConnect.dataConnectPath(),
         Arb.enum<Storage>(),
       )
-    checkAll(propTestConfig, cacheSettingsArb(), otherArb) {
+    checkAll(propTestConfig, Arb.dataConnect.cacheSettings(), otherArb) {
       cacheSettings: CacheSettings,
       other: Any ->
       cacheSettings.equals(other) shouldBe false
@@ -129,7 +128,7 @@ class CacheSettingsUnitTest {
   @Test
   fun `hashCode() should return the same value each time it is invoked on a given object`() =
     runTest {
-      checkAll(propTestConfig, cacheSettingsArb()) { cacheSettings: CacheSettings ->
+      checkAll(propTestConfig, Arb.dataConnect.cacheSettings()) { cacheSettings: CacheSettings ->
         val hashCode1 = cacheSettings.hashCode()
         cacheSettings.hashCode() shouldBe hashCode1
         cacheSettings.hashCode() shouldBe hashCode1
@@ -138,7 +137,7 @@ class CacheSettingsUnitTest {
 
   @Test
   fun `hashCode() should return the same value on equal objects`() = runTest {
-    checkAll(propTestConfig, cacheSettingsArb()) { cacheSettings1: CacheSettings ->
+    checkAll(propTestConfig, Arb.dataConnect.cacheSettings()) { cacheSettings1: CacheSettings ->
       val cacheSettings2 = cacheSettings1.copy()
       withClue("cacheSettings1=$cacheSettings1 cacheSettings2=$cacheSettings2") {
         cacheSettings1.hashCode() shouldBe cacheSettings2.hashCode()
@@ -159,7 +158,7 @@ class CacheSettingsUnitTest {
 
   @Test
   fun `copy with no arguments should create a distinct, but equal object`() = runTest {
-    checkAll(propTestConfig, cacheSettingsArb()) { cacheSettings1: CacheSettings ->
+    checkAll(propTestConfig, Arb.dataConnect.cacheSettings()) { cacheSettings1: CacheSettings ->
       val cacheSettings2 = cacheSettings1.copy()
       withClue("cacheSettings1=$cacheSettings1 cacheSettings2=$cacheSettings2") {
         assertSoftly {
@@ -192,8 +191,3 @@ class CacheSettingsUnitTest {
       )
   }
 }
-
-// TODO: Move to cacheSettingsArb() once cache api goes public
-private fun cacheSettingsArb(
-  storage: Arb<Storage> = Arb.enum<Storage>(),
-): Arb<CacheSettings> = storage.map(::CacheSettings)